refactor(importer): extract note formatter concepts
Jira: AI-884
