+++ app/lib/customs_tariff_importer/notes_extractor/formatter.rb
@@ -1,193 +1,6 @@
 module CustomsTariffImporter
   class NotesExtractor
     class Formatter
-      class ParagraphPreprocessor
-        COMPACT_NUMBERED_MARKER_PATTERN = /\A([1-9]\d*)\.(?=(?:\([a-z]\)|[A-Z]\.|[[:alpha:]]))/i
-        DOTLESS_NUMBERED_NOTE_PATTERN = /\A([1-9]\d*)\s+(?=[A-Z]|\u201C|\()/
-        NUMBERED_NOTE_PATTERN = /\A[1-9]\d*\.(?!\d)/
-        SOURCE_BULLET_MARKER_PATTERN = /\A\s*(?:-|•|—|–)\s*/
-
-        def initialize(numbering_counters)
-          @numbering_counters = numbering_counters
-        end
-
-        def call(text, current_paragraph:, last_top_level_note_number:)
-          text = normalize_source_bullet_marker(text)
-          text = normalize_compact_numbered_marker(text)
-          text = normalize_dotless_numbered_note(text, last_top_level_note_number)
-          normalize_visual_numbering(text, current_paragraph)
-        end
-
-        private
-
-        def normalize_source_bullet_marker(text)
-          return text unless text.match?(SOURCE_BULLET_MARKER_PATTERN)
-
-          "- #{text.sub(SOURCE_BULLET_MARKER_PATTERN, '')}"
-        end
-
-        def normalize_compact_numbered_marker(text)
-          text.sub(COMPACT_NUMBERED_MARKER_PATTERN, '\1. ')
-        end
-
-        def normalize_dotless_numbered_note(text, last_top_level_note_number)
-          return text unless (match = text.match(DOTLESS_NUMBERED_NOTE_PATTERN))
-
-          note_number = match[1].to_i
-          expected_note_number = last_top_level_note_number ? last_top_level_note_number + 1 : 1
-          continues_sequence = note_number == expected_note_number
-          return text unless continues_sequence
-
-          text.sub(/\A([1-9]\d*)\s+/, '\1. ')
-        end
-
-        def normalize_visual_numbering(text, current_paragraph)
-          numbering = current_paragraph&.fetch(:numbering, nil)
-          return text unless numbering && numbering[:id].positive? && numbering[:level].zero?
-          return text if text.match?(NUMBERED_NOTE_PATTERN)
-
-          "#{next_visual_number(numbering)}. #{text}"
-        end
-
-        def next_visual_number(numbering)
-          key = [numbering[:id], numbering[:level]]
-          @numbering_counters[key] = @numbering_counters.fetch(key, 0) + 1
-        end
-      end
-
-      class MarkerTrie
-        ROMAN_MARKER_PATTERN = /(?:i{1,3}|iv|v|vi{0,3}|ix)/i
-        ALPHABETIC_DOTTED_MARKER_PATTERN = /(?:[a-z]|ij)\./i
-        MARKER_TEXT_PATTERN = /\A(?:#{ALPHABETIC_DOTTED_MARKER_PATTERN}|\([A-Z]\)|\([ivxlcdm]+\)|[ivxlcdm]+\.|\(\d+\)|\d+\.)\s+\S/i
-        ALPHABETIC_MARKER_TEXT_PATTERN = /\A#{ALPHABETIC_DOTTED_MARKER_PATTERN}\s+\S/i
-        ROMAN_MARKER_TEXT_PATTERN = /\A#{ROMAN_MARKER_PATTERN}\.\s+\S/
-        CHILD_MARKER_TEXT_PATTERN = /\A(?:\(\d+\)|\d+\.|#{ROMAN_MARKER_PATTERN}\.|\(#{ROMAN_MARKER_PATTERN}\))\s+\S/
-        SOURCE_MARKER_LINE_PATTERN = /\A\s+(?:-\s+)?(?:#{ALPHABETIC_DOTTED_MARKER_PATTERN}|\([A-Z]\)|\([ivxlcdm]+\)|[ivxlcdm]+\.|\(\d+\)|\d+\.)\s+\S/i
-        MARKER_LINE_PATTERN = /\A(?<spaces> *)(?:-\s+)?(?<marker>\(\d+\)|\d+\.|ij\.|#{ROMAN_MARKER_PATTERN}\.|[a-z]\.|\([a-z]+\))\s+\S/i
-        SOURCE_PARENT_MARKER_LINE_PATTERN = /\A\s+(?:-\s+)?(?:#{ALPHABETIC_DOTTED_MARKER_PATTERN}|\((?![ivx]+\))[a-z]\))\s+\S/i
-        SOURCE_CHILD_MARKER_LINE_PATTERN = /\A\s+(?:-\s+)?(?:\(\d+\)|\d+\.|#{ROMAN_MARKER_PATTERN}\.|\(#{ROMAN_MARKER_PATTERN}\))\s+\S/
-
-        def initialize
-          reset
-        end
-
-        def reset
-          @pending_parent_depth = nil
-          @active_child_family = nil
-        end
-
-        def child_indent_level(text)
-          return unless @pending_parent_depth
-
-          marker = self.class.marker_from_text(text)
-          return unless marker && self.class.child_marker?(marker)
-
-          family = self.class.child_marker_family(marker)
-          if @active_child_family
-            @pending_parent_depth + 1 if family == @active_child_family
-          elsif self.class.first_child_marker?(marker)
-            @pending_parent_depth + 1
-          end
-        end
-
-        def observe(formatted, in_numbered_note:, paragraph_indent_level:, paragraph_first_line_indent_level:)
-          if (paragraph_indent_level.positive? && formatted.match?(SOURCE_CHILD_MARKER_LINE_PATTERN)) ||
-              (paragraph_indent_level.zero? && paragraph_first_line_indent_level.positive? && self.class.marker_line?(formatted, :child))
-            marker = self.class.marker_from_line(formatted)
-            @active_child_family = self.class.child_marker_family(marker) if marker
-            @pending_parent_depth
-          elsif (paragraph_indent_level.positive? && formatted.match?(SOURCE_PARENT_MARKER_LINE_PATTERN)) ||
-              (paragraph_indent_level.zero? && paragraph_first_line_indent_level.positive? && self.class.marker_line?(formatted, :parent)) ||
-              (in_numbered_note && self.class.marker_line?(formatted, :parent))
-            @pending_parent_depth = formatted_indent_level(formatted)
-            @active_child_family = nil
-          else
-            reset
-          end
-        end
-
-        def self.source_marker_line?(line)
-          line.to_s.match?(SOURCE_MARKER_LINE_PATTERN)
-        end
-
-        def self.marker_text?(text)
-          text.match?(MARKER_TEXT_PATTERN)
-        end
-
-        def self.alphabetic_marker_text?(text)
-          text.match?(ALPHABETIC_MARKER_TEXT_PATTERN)
-        end
-
-        def self.roman_marker_text?(text)
-          text.match?(ROMAN_MARKER_TEXT_PATTERN)
-        end
-
-        def self.marker_line?(line, kind, min_indent: 1)
-          parsed_line(line)&.then do |marker|
-            marker[:indent_level] >= min_indent && public_send("#{kind}_marker?", marker[:text])
-          end
-        end
-
-        def self.indent_level(line)
-          parsed_line(line)&.fetch(:indent_level)
-        end
-
-        def self.marker_from_text(text)
-          text.to_s.match(MARKER_LINE_PATTERN)&.[](:marker)
-        end
-
-        def self.marker_from_line(line)
-          parsed_line(line)&.fetch(:text)
-        end
-
-        def self.parsed_line(line)
-          return unless (match = line.to_s.match(MARKER_LINE_PATTERN))
-
-          { indent_level: match[:spaces].length / 4, text: match[:marker] }
-        end
-
-        def self.parent_marker?(marker)
-          (marker.match?(/\A#{ALPHABETIC_DOTTED_MARKER_PATTERN}\z/i) && !roman_child_marker?(marker)) ||
-            marker.match?(/\A\((?![ivx]+\))[a-z]+\)\z/i)
-        end
-
-        def self.first_parent_marker?(marker)
-          marker.match?(/\Aa\.\z/) || marker.match?(/\A\(a\)\z/)
-        end
-
-        def self.promotable_parent_marker?(marker)
-          first_parent_marker?(marker) || marker.match?(/\A\(A\)\z/)
-        end
-
-        def self.uppercase_parent_marker?(marker)
-          marker.match?(/\A\([A-Z]\)\z/)
-        end
-
-        def self.child_marker?(marker)
-          marker.match?(/\A(?:\d+\.|\(\d+\)|#{ROMAN_MARKER_PATTERN}\.|\(#{ROMAN_MARKER_PATTERN}\))\z/)
-        end
-
-        def self.roman_child_marker?(marker)
-          marker.match?(/\A(?:#{ROMAN_MARKER_PATTERN}\.|\(#{ROMAN_MARKER_PATTERN}\))\z/i)
-        end
-
-        def self.first_child_marker?(marker)
-          marker.match?(/\A(?:1\.|\(1\)|i\.|\(i\))\z/i)
-        end
-
-        def self.child_marker_family(marker)
-          marker.match?(/\A(?:\d+\.|\(\d+\))\z/) ? :number : :roman
-        end
-
-        private_class_method :parsed_line
-
-        private
-
-        def formatted_indent_level(formatted)
-          formatted[/\A */].length / 4
-        end
-      end
-
       def initialize
         reset_note_formatting_context
       end
@@ -225,6 +38,7 @@ module CustomsTariffImporter
         @numbering_counters = {}
         @paragraph_preprocessor = ParagraphPreprocessor.new(@numbering_counters)
         @marker_trie = MarkerTrie.new
+        @table_formatter = TableFormatter.new
       end
 
       def normalize_note_lines(lines, chapter: false, section: false, current_chapter: nil)
@@ -332,137 +146,7 @@ module CustomsTariffImporter
       end
 
       def table_markdown_lines(table)
-        rows = table.xpath('./w:tr', WORD_NS).map { |row| table_row_cells(row) }
-        rows.reject!(&:empty?)
-        return [] if rows.empty?
-        return [rows.flatten.find(&:present?)] if commodity_table?(rows)
-        return grouped_table_markdown_lines(rows) if grouped_table_header?(rows)
-
-        header, body_rows = table_header_and_body(rows)
-        body = body_rows.map { |row| normalize_table_row(row, header.length) }
-        ['', markdown_table_row(header), markdown_table_separator(header.length), *body.map { |row| markdown_table_row(row) }, '']
-      end
-
-      def table_row_cells(row)
-        row.xpath('./w:tc', WORD_NS).flat_map do |cell|
-          text = cell.xpath('./w:p', WORD_NS)
-            .map { |para| paragraph_plain_text(para) }
-            .reject(&:blank?)
-            .join(' ')
-
-          Array.new(table_cell_span(cell), text)
-        end
-      end
-
-      def table_cell_span(cell)
-        cell.at_xpath('./w:tcPr/w:gridSpan', WORD_NS)&.attr('w:val').to_i.then { |span| span.positive? ? span : 1 }
-      end
-
-      def commodity_table?(rows)
-        rows.flatten.find(&:present?).to_s.match?(COMMODITY_CODE_PATTERN)
-      end
-
-      def table_header_and_body(rows)
-        header_rows = table_header_rows(rows)
-        header = combine_table_header_rows(header_rows)
-
-        [header, rows.drop(header_rows.length)]
-      end
-
-      def grouped_table_markdown_lines(rows)
-        columns = rows.map(&:length).max
-        header = collapse_spanned_cells(normalize_table_row(rows.first, columns))
-        markers = collapse_spanned_cells(normalize_table_row(rows.second, columns))
-        subheader = normalize_table_row(rows.third, columns)
-        body = rows.drop(3).map { |row| normalize_table_row(row, columns) }
-
-        [
-          '',
-          markdown_table_row(header),
-          markdown_table_separator(columns),
-          markdown_table_row(markers),
-          markdown_table_row(subheader),
-          *body.map { |row| markdown_table_row(row) },
-          '',
-        ]
-      end
-
-      def collapse_spanned_cells(row)
-        previous = nil
-        row.map do |cell|
-          collapsed = cell == previous ? '' : cell
-          previous = cell
-          collapsed
-        end
-      end
-
-      def grouped_table_header?(rows)
-        return false if rows.length < 4
-        return false unless rows.first.length == rows.third.length
-        return false unless rows.first.chunk_while { |left, right| left == right }.any? { |chunk| chunk.length > 1 }
-        return false unless rows.second.compact_blank.all? { |cell| cell.match?(/\A\(\d+\)\z/) }
-
-        rows.third.any?(&:blank?) && rows.third.any?(&:present?)
-      end
-
-      def table_header_rows(rows)
-        return rows.first(2) if multi_row_header?(rows)
-
-        [collapse_duplicate_header_cells(rows.first)]
-      end
-
-      def collapse_duplicate_header_cells(row)
-        row.chunk_while { |left, right| left == right }.map(&:first)
-      end
-
-      def multi_row_header?(rows)
-        return false if rows.length < 3
-        return false unless rows.second.any?(&:blank?)
-
-        rows.first.zip(rows.second).any? { |top, lower| top.present? && lower.present? }
-      end
-
-      def combine_table_header_rows(rows)
-        columns = rows.map(&:length).max
-        normalized = rows.map { |row| normalize_table_row(row, columns) }
-
-        Array.new(columns) do |index|
-          normalized.map { |row| row[index] }.reject(&:blank?).uniq.join(' ')
-        end
-      end
-
-      def normalize_table_row(row, columns)
-        if columns == 2 && row.length == 3
-          [combined_two_column_table_label(row.first(2)), row.last]
-        else
-          row.first(columns).fill('', row.length...columns)
-        end
-      end
-
-      def combined_two_column_table_label(cells)
-        compact_cells = cells.reject(&:blank?)
-        return compact_cells.first.to_s if compact_cells.uniq.one?
-        return "#{compact_cells.first} (#{compact_cells.second})" if chemical_symbol_name_cells?(compact_cells)
-
-        compact_cells.join(' ')
-      end
-
-      def chemical_symbol_name_cells?(cells)
-        cells.length == 2 &&
-          cells.first.match?(/\A[A-Z][a-z]?\z/) &&
-          cells.second.match?(/\A[A-Z][a-z]+\z/)
-      end
-
-      def markdown_table_row(cells)
-        "| #{cells.map { |cell| markdown_table_cell(cell) }.join(' | ')} |"
-      end
-
-      def markdown_table_separator(columns)
-        markdown_table_row(Array.new(columns, '---'))
-      end
-
-      def markdown_table_cell(text)
-        text.to_s.strip.gsub(/[\\|]/) { |character| "\\#{character}" }
+        @table_formatter.call(table)
       end
 
       def formatted_note_line(text)
